requirements.txt
Flask==1.1.2
python-dateutil==2.8.1
pytz==2021.1
uWSGI==2.0.19.1
python-dotenv==0.15.0
